Separate auth to helper function
